Move files to the under acctf and fix requirements
