+++ acctf/utils/format.py
