+++ requirements.txt
@@ -1,7 +1,7 @@
-beautifulsoup4==4.12.2
-bs4==0.0.1
-chromedriver-binary==122.0.6226.0.0
-lxml==5.0.0
-pandas==2.1.4
-PySocks==1.7.1
-selenium==4.16.0
+beautifulsoup4>=4.12.2
+bs4>=0.0.1
+chromedriver-binary>=122.0.6226.0.0
+lxml>=5.0.0
+pandas>=2.1.4
+PySocks>=1.7.1
+selenium>=4.16.0
